requirements.txt
python-dotenv
pandas
langchain-groq
langchain-ollama
langchain-core
langgraph
langgraph-supervisor
typing-extensions
pydantic
fastapi
uvicorn
streamlit
requests
- e.